1.42 new fan speed control

+++ Repetier/Printer.cpp
@@ -407,7 +407,7 @@ void Printer::kill(uint8_t only_steppers)
 
 #if FAN_PIN>-1 && FEATURE_FAN_CONTROL
     // disable the fan
-    Commands::setFanSpeed(0);
+    Commands::setFanSpeed((uint8_t)0);
 #endif // FAN_PIN>-1 && FEATURE_FAN_CONTROL
 
     disableAllSteppersNow();
